+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-2.5, 2024-06-18
+2.5, 2024-06-19
 ---------------
 - add save-button to individual heroes
 - add buttons to set hero level from experience and vice versa
@@ -14,6 +14,7 @@ CHANGELOG
 - cap hero level at 75 (74 in Horn of the Abyss)
 - cap hero level at 2^31 - 1 instead of 2^32 - 1
 - show different error on trying to open files no longer on disk
+- use current directory and extension filter in file dialogs
 
 
 2.4, 2024-06-14